renamed YoutubeVideo component
